requirements.txt
pinecone==6.0.2
google-generativeai==0.8.4
langchain==0.3.21
langchain-google-genai==2.1.2
python-dotenv==1.1.0
fastapi==0.115.12
uvicorn==0.34.0
requests==2.31.0
